+++ CHANGES.rst
@@ -37,6 +37,7 @@ Changelog
     `MeetingConfig.useGroupsAsCategories` was `True`, now it returns only categories, moreover parameter `classifiers` is
     renamed to `catType` that may be `all`/`categories`/`classifiers`.
 - In every migrations, call `cleanRegistries` at the end by default so `JS/CSS` are recompiled
+- Make sure `++resource++plone.app.jquerytools.dateinput.js` is enabled in `portal_javascripts`
 
 4.1.27.2 (2020-06-25)
 ---------------------