+++ CHANGES.rst
@@ -2,7 +2,8 @@ Changelog
 =========
 
 
-4.1.28 (unreleased) 
+
+4.1.28 (unreleased)
 -------------------
 
 - Moved `Meeting.getNextMeeting` logic to `utils.get_next_meeting` so it can be used from outside a `Meeting` instance,
@@ -37,10 +38,9 @@ Changelog
     `MeetingConfig.useGroupsAsCategories` was `True`, now it returns only categories, moreover parameter `classifiers` is
     renamed to `catType` that may be `all`/`categories`/`classifiers`.
 - In every migrations, call `cleanRegistries` at the end by default so `JS/CSS` are recompiled
-
 - Add 'redirectToNextMeeting' option.
-
 - Moved `Meeting.getNextMeeting` logic to `utils.get_next_meeting` so it can be used from outside a `Meeting` instance
+- Make sure `++resource++plone.app.jquerytools.dateinput.js` is enabled in `portal_javascripts`
 
 
 4.1.27.2 (2020-06-25)